renamed gizmo files to follow convention
