+++ extras/gizmo/scale-gizmo.js
@@ -1,24 +1,27 @@
 import {
-    Quat,
     Vec3
 } from 'playcanvas';
 
-import { AxisArrow, AxisPlane } from './axis-shapes.js';
+import { AxisBoxCenter, AxisBoxLine, AxisPlane } from './axis-shapes.js';
 import { LOCAL_COORD_SPACE } from './gizmo.js';
-import { GizmoTransform } from "./gizmo-transform.js";
+import { TransformGizmo } from "./transform-gizmo.js";
 
 // temporary variables
 const tmpV1 = new Vec3();
-const tmpV2 = new Vec3();
-const tmpQ1 = new Quat();
 
 /**
- * Translation gizmo.
+ * Scaling gizmo.
  *
- * @augments GizmoTransform
+ * @augments TransformGizmo
  */
-class GizmoTranslate extends GizmoTransform {
+class ScaleGizmo extends TransformGizmo {
     _shapes = {
+        xyz: new AxisBoxCenter(this.app.graphicsDevice, {
+            axis: 'xyz',
+            layers: [this.layer.id],
+            defaultColor: this._materials.axis.xyz,
+            hoverColor: this._materials.hover.xyz
+        }),
         yz: new AxisPlane(this.app.graphicsDevice, {
             axis: 'x',
             flipAxis: 'y',
@@ -43,21 +46,21 @@ class GizmoTranslate extends GizmoTransform {
             defaultColor: this._materials.axis.z.cullNone,
             hoverColor: this._materials.hover.z.cullNone
         }),
-        x: new AxisArrow(this.app.graphicsDevice, {
+        x: new AxisBoxLine(this.app.graphicsDevice, {
             axis: 'x',
             layers: [this.layer.id],
             rotation: new Vec3(0, 0, -90),
             defaultColor: this._materials.axis.x.cullBack,
             hoverColor: this._materials.hover.x.cullBack
         }),
-        y: new AxisArrow(this.app.graphicsDevice, {
+        y: new AxisBoxLine(this.app.graphicsDevice, {
             axis: 'y',
             layers: [this.layer.id],
             rotation: new Vec3(0, 0, 0),
             defaultColor: this._materials.axis.y.cullBack,
             hoverColor: this._materials.hover.y.cullBack
         }),
-        z: new AxisArrow(this.app.graphicsDevice, {
+        z: new AxisBoxLine(this.app.graphicsDevice, {
             axis: 'z',
             layers: [this.layer.id],
             rotation: new Vec3(90, 0, 0),
@@ -66,55 +69,82 @@ class GizmoTranslate extends GizmoTransform {
         })
     };
 
-    /**
-     * Internal mapping from each attached node to their starting position in local space.
-     *
-     * @type {Map<import('playcanvas').GraphNode, Vec3>}
-     * @private
-     */
-    _nodeLocalPositions = new Map();
+    _coordSpace = LOCAL_COORD_SPACE;
 
     /**
-     * Internal mapping from each attached node to their starting position in world space.
+     * Internal mapping from each attached node to their starting scale.
      *
      * @type {Map<import('playcanvas').GraphNode, Vec3>}
      * @private
      */
-    _nodePositions = new Map();
+    _nodeScales = new Map();
+
+    /**
+     * State for if uniform scaling is enabled for planes. Defaults to true.
+     *
+     * @type {boolean}
+     */
+    uniform = true;
 
     snapIncrement = 1;
 
     /**
-     * Creates a new GizmoTranslate object.
+     * Creates a new ScaleGizmo object.
      *
      * @param {import('playcanvas').AppBase} app - The application instance.
      * @param {import('playcanvas').CameraComponent} camera - The camera component.
      * @param {import('playcanvas').Layer} layer - The render layer.
      * @example
-     * const gizmo = new pcx.GizmoTranslate(app, camera, layer);
+     * const gizmo = new pcx.ScaleGizmo(app, camera, layer);
      */
     constructor(app, camera, layer) {
         super(app, camera, layer);
 
         this._createTransform();
 
+        this.on('key:down', (key, shiftKey, ctrlKey) => {
+            this.uniform = ctrlKey;
+        });
+
+        this.on('key:up', () => {
+            this.uniform = false;
+        });
+
         this.on('transform:start', () => {
-            this._storeNodePositions();
+            this._selectionStartPoint.sub(Vec3.ONE);
+            this._storeNodeScales();
         });
 
         this.on('transform:move', (pointDelta) => {
+            const axis = this._selectedAxis;
+            const isPlane = this._selectedIsPlane;
             if (this.snap) {
                 pointDelta.scale(1 / this.snapIncrement);
                 pointDelta.round();
                 pointDelta.scale(this.snapIncrement);
             }
-            this._setNodePositions(pointDelta);
+            if (this.uniform && isPlane) {
+                tmpV1.set(Math.abs(pointDelta.x), Math.abs(pointDelta.y), Math.abs(pointDelta.z));
+                tmpV1[axis] = 0;
+                const v = tmpV1.length();
+                tmpV1.set(v * Math.sign(pointDelta.x), v * Math.sign(pointDelta.y), v * Math.sign(pointDelta.z));
+                tmpV1[axis] = 1;
+                pointDelta.copy(tmpV1);
+            }
+            this._setNodeScales(pointDelta);
         });
 
         this.on('nodes:detach', () => {
-            this._nodeLocalPositions.clear();
-            this._nodePositions.clear();
+            this._nodeScales.clear();
         });
+    }
+
+    set coordSpace(value) {
+        // disallow changing coordSpace for scale
+    }
+
+    get coordSpace() {
+        return this._coordSpace;
     }
 
     set axisGap(value) {
@@ -149,20 +179,12 @@ class GizmoTranslate extends GizmoTransform {
         return this._shapes.x.tolerance;
     }
 
-    set axisArrowThickness(value) {
-        this._setArrowProp('arrowThickness', value);
+    set axisBoxSize(value) {
+        this._setArrowProp('boxSize', value);
     }
 
-    get axisArrowThickness() {
-        return this._shapes.x.arrowThickness;
-    }
-
-    set axisArrowLength(value) {
-        this._setArrowProp('arrowLength', value);
-    }
-
-    get axisArrowLength() {
-        return this._shapes.x.arrowLength;
+    get axisBoxSize() {
+        return this._shapes.x.boxSize;
     }
 
     set axisPlaneSize(value) {
@@ -181,6 +203,23 @@ class GizmoTranslate extends GizmoTransform {
         return this._shapes.yz.gap;
     }
 
+    set axisCenterSize(value) {
+        this._shapes.xyz.size = value;
+    }
+
+    get axisCenterSize() {
+        return this._shapes.xyz.size;
+    }
+
+    set axisCenterTolerance(value) {
+        this._shapes.xyz.tolerance = value;
+    }
+
+    get axisCenterTolerance() {
+        return this._shapes.xyz.tolerance;
+    }
+
+
     _setArrowProp(prop, value) {
         this._shapes.x[prop] = value;
         this._shapes.y[prop] = value;
@@ -193,33 +232,19 @@ class GizmoTranslate extends GizmoTransform {
         this._shapes.xy[prop] = value;
     }
 
-    _storeNodePositions() {
+    _storeNodeScales() {
         for (let i = 0; i < this.nodes.length; i++) {
             const node = this.nodes[i];
-            this._nodeLocalPositions.set(node, node.getLocalPosition().clone());
-            this._nodePositions.set(node, node.getPosition().clone());
+            this._nodeScales.set(node, node.getLocalScale().clone());
         }
     }
 
-    _setNodePositions(pointDelta) {
+    _setNodeScales(pointDelta) {
         for (let i = 0; i < this.nodes.length; i++) {
             const node = this.nodes[i];
-            if (this._coordSpace === LOCAL_COORD_SPACE) {
-                tmpV1.copy(pointDelta);
-                node.parent.getWorldTransform().getScale(tmpV2);
-                tmpV2.x = 1 / tmpV2.x;
-                tmpV2.y = 1 / tmpV2.y;
-                tmpV2.z = 1 / tmpV2.z;
-                tmpQ1.copy(node.getLocalRotation()).transformVector(tmpV1, tmpV1);
-                tmpV1.mul(tmpV2);
-                node.setLocalPosition(this._nodeLocalPositions.get(node).clone().add(tmpV1));
-            } else {
-                node.setPosition(this._nodePositions.get(node).clone().add(pointDelta));
-            }
+            node.setLocalScale(this._nodeScales.get(node).clone().mul(pointDelta));
         }
-
-        this._updatePosition();
     }
 }
 
-export { GizmoTranslate };
+export { ScaleGizmo };
