+++ extras/index.js
@@ -28,7 +28,7 @@ export { RenderPassTAA } from './render-passes/render-pass-taa.js';
 
 // gizmo
 export { Gizmo } from "./gizmo/gizmo.js";
-export { GizmoTransform } from "./gizmo/gizmo-transform.js";
-export { GizmoTranslate } from "./gizmo/gizmo-translate.js";
-export { GizmoRotate } from "./gizmo/gizmo-rotate.js";
-export { GizmoScale } from "./gizmo/gizmo-scale.js";
+export { TransformGizmo } from "./gizmo/transform-gizmo.js";
+export { TranslateGizmo } from "./gizmo/translate-gizmo.js";
+export { RotateGizmo } from "./gizmo/rotate-gizmo.js";
+export { ScaleGizmo } from "./gizmo/gizmo-scale.js";
